move deep merge to public submodule
