+++ hydromt/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Public helper functions that (pulgin) developers can use."""
+
+from hydromt.utils.deep_merge import deep_merge
+
+__all__ = ["deep_merge"]
