+++ tests/data_sources/test_geo_data_frame_source.py
@@ -49,7 +49,7 @@ class TestGeoDataFrameSource:
         error_driver = next(
             filter(lambda e: e["loc"] == ("driver",), e_info.value.errors())
         )
-        assert error_driver["type"] == "modeltype"
+        assert error_driver["type"] == "model_type"
 
     def test_raises_on_invalid_fields(
         self,
